pyproject.toml
# ==========================
# Main project configuration 
# ==========================

[project]
name = "djresttoolkit"
version = "1.1.0"
description = "A collection of Django and DRF utilities to simplify API development."
readme = { file = "README.md", content-type = "text/markdown" }
license = { file = "LICENSE" }
authors = [
    { name = "shaileshpandit141", email = "shaileshpandit141@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
]
keywords = [
    "django",
    "djangorestframework",
    "drf",
    "django-utilities",
    "django-rest-utilities",
    "django-rest-framework-utilities",
    "django-helpers",
    "django-rest",
    "rest-api",
    "api-development",
    "django-email",
    "django-authentication",
    "django-serializers",
    "django-pagination",
    "django-mixins",
    "django-shortcuts",
    "python-package",
]
requires-python = ">=3.13"
dependencies = ["faker>=37.5.3", "pydantic>=2.11.7"]

# CLI scripts entry point configuration
# -------------------------------------
[project.scripts]
djresttoolkit = "djresttoolkit:entry_point"

# ==========================
# Build system configuration
# ==========================
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# =================
# Dependency groups
# =================
[dependency-groups]
dev = [
    "django>=5.2.5",
    "django-stubs>=5.2.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs>=3.16.2",
    "pytest>=8.4.1",
    "djresttoolkit",
    "pyyaml>=6.0.2",
    "pydantic-settings>=2.10.1",
    "daphne>=4.2.1",
    "django-redis>=6.0.0",
    "django-cors-headers>=4.7.0",
    "argon2-cffi-bindings>=21.2.0",
]

# =====================
# Optional dependencies
# =====================
[project.optional-dependencies]
dev = ["pytest", "mypy", "ruff"]

# ===============
# UV tool sources
# ===============
[tool.uv.sources]
djresttoolkit = { workspace = true }

# ============
# Package data
# ============
[tool.setuptools.package-data]
"djresttoolkit" = ["py.typed"]

# ============
# Project URLs
# ============
[project.urls]
Homepage = "https://github.com/shaileshpandit141/djresttoolkit"
Documentation = "https://shaileshpandit141.github.io/djresttoolkit"
Source = "https://github.com/shaileshpandit141/djresttoolkit"
Issues = "https://github.com/shaileshpandit141/djresttoolkit/issues"
License = "https://github.com/shaileshpandit141/djresttoolkit/blob/main/LICENSE"


# ==========================
# Add pyright Configurations
# ==========================
[tool.pyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportIncompatibleVariableOverride = false

# =========================================
# Disable ruff S101 rule for test casecases
# =========================================
[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

# ====================
# Pytest configuration
# ====================
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
